feat: Add session cookie as possible URL parameter
This makes it possible to use the CCM with the xpra desktop client
